JBIDE-12175 Automatic namespace completion in Facelets files
Implemented name space storage for user defined namespaces.
Automatic namespace completion is added for the namespaces that are not
defined within the page:
- xmlns:*-attribute names
- xmlns:*-attribute values
- tags having a prefix that is not defined within the page
All the default prefixes of the Tag Libraries and the custom user defined prefixes are
taken into account.
JUnit Test is added due to verify the issue:
-  tests/org.jboss.tools.jst.jsp.test/src/org/jboss/tools/jst/jsp/test/ca/JstCAOnCustomPrefixesTest.java