JBIDE-13864 Memory usage problem in PageContextFactory.CSSStyleSheetDescriptor
The loading of CSS Stylesheet Descriptors is excluded from Page Context Creation. CSS Stylesheet Descriptors
are loaded in on-demand manner when they are requested by Content Assistant or Hyperlink features.
As such Page Contexts don't store a links to DOM Nodes reducing the memory overload and are loaded a little bit faster.

	modified:   plugins/org.jboss.tools.jst.web.kb/src/org/jboss/tools/jst/web/kb/PageContextFactory.java
	modified:   plugins/org.jboss.tools.jst.web.kb/src/org/jboss/tools/jst/web/kb/internal/JspContextImpl.java