JBIDE-14597 warning about ng-* attributes used by angular js should be possible to ignore
The UI interface is added for HTML5 documents that allows to ignore in validation attribute names specified by a user:
Preference Page, Project Property page and QuickFix are added to allow a user to enable such a facility and to specify
the names of attributes to be ignored by HTMLAttributeValidator
The fix is made with respect of WTP version less than 3.5.2SR2 that have no required preferences support. The pref. page and
quick fix are still showing on the older platforms but do not perform any changes. The only indicatition is displayed to make a user
to update to the latest WTP 3.5.2SR2.
JUnit Test is added due to test the issue

Signed-off-by: vrubezhny <vrubezhny@exadel.com>